More Tk stuff, May 24, 2016

+++ src/tk/generic/tkTextImage.c
@@ -374,7 +374,7 @@ EmbImageConfigure(textPtr, eiPtr, argc, argv)
 	    if (new > count) {
 		count = new;
 	    }
-	    if (len == (int) strlen(haveName)) {
+	    if (len == strlen(haveName)) {
 	    	conflict = 1;
 	    }
 	}
